Refactor text util tests and remove unused code
