+++ utils/text.py
@@ -20,11 +20,13 @@
 
 import re
 import unicodedata
-import bleach
-from bleach.html5lib_shim import Filter
 from functools import partial
 from htmlentitydefs import name2codepoint
+
+import bleach
+from bleach.html5lib_shim import Filter
 from django.utils.encoding import smart_unicode
+
 from sounds.templatetags.sound_signature import SOUND_SIGNATURE_SOUND_ID_PLACEHOLDER, \
     SOUND_SIGNATURE_SOUND_URL_PLACEHOLDER
 
@@ -35,19 +37,19 @@ def slugify(s, entities=True, decimal=True, hexadecimal=True, instance=None, slu
 
     #  character entity reference
     if entities:
-        s = re.sub('&(%s);' % '|'.join(name2codepoint), lambda m: unichr(name2codepoint[m.group(1)]), s)
+        s = re.sub(r'&(%s);' % '|'.join(name2codepoint), lambda m: unichr(name2codepoint[m.group(1)]), s)
 
     #  decimal character reference
     if decimal:
         try:
-            s = re.sub('&#(\d+);', lambda m: unichr(int(m.group(1))), s)
+            s = re.sub(r'&#(\d+);', lambda m: unichr(int(m.group(1))), s)
         except:
             pass
 
     #  hexadecimal character reference
     if hexadecimal:
         try:
-            s = re.sub('&#x([\da-fA-F]+);', lambda m: unichr(int(m.group(1), 16)), s)
+            s = re.sub(r'&#x([\da-fA-F]+);', lambda m: unichr(int(m.group(1), 16)), s)
         except:
             pass
 
@@ -86,57 +88,13 @@ def shout_percentage(string):
 
 
 def is_shouting(string):
-    """
-    >>> is_shouting('')
-    False
-    >>> is_shouting('HELLO THIS IS SHOUTING!!!')
-    True
-    >>> is_shouting('This is a phrase WITH SONE emphasis!!')
-    False
-    >>> is_shouting('This is a regular phrase.')
-    False
-    """
+    """Test if a string sounds like shouting (more than 60% uppercase)"""
     if len(string) < 5:
         return False
     return shout_percentage(string) > 0.6
 
 
-# returns if the string ends with any of the endinggs
-def endswithone(string, endings):
-    reduce(lambda a,b: a or string.endswith(b), endings, False)
-
-
-# returns if the string starts with any of the starts
-def startswithone(string, starts):
-    reduce(lambda a,b: a or string.startswith(b), starts, False)
-
-
-url_regex = re.compile("(https?://\S+)", re.IGNORECASE)
-
-
-def replace_element_by_children(soup, element):
-    """
-    replace an element in the DOM with it's child nodes
-    """
-    parent = element.parent
-
-    if not parent:
-        parent = soup
-
-    # afterwards we need to insert the children where the parent used to be!
-    position_in_parent = 0
-    for c in parent.contents:
-        if c == element:
-            break
-        else:
-            position_in_parent += 1
-
-    # in reverse order, insert the child elements in their place.
-    for el in element.contents[::-1]:
-        parent.insert(position_in_parent, el)
-
-    # delete the element
-    element.extract()
+url_regex = re.compile(r"(https?://\S+)", re.IGNORECASE)
 
 
 def nofollow(attrs, new=False):
@@ -167,8 +125,9 @@ class EmptyLinkFilter(Filter):
                     continue
             yield token
 
+
 def clean_html(input):
-    # Reaplce html tags from user input, see utils.test for examples
+    # Replace html tags from user input, see utils.test for examples
 
     ok_tags = [u"a", u"img", u"strong", u"b", u"em", u"i", u"u", u"ul", u"li", u"p", u"br",  u"blockquote", u"code"]
     ok_attributes = {u"a": [u"href", u"rel"], u"img": [u"src", u"alt", u"title"]}
@@ -189,44 +148,16 @@ def clean_html(input):
     output = cleaner.clean(input)
     return output
 
+
 def remove_control_chars(text):
-    return ''.join(c for c in text if (ord(c) >= 32 or ord(c) in [9,10,13]))
+    return ''.join(c for c in text if (ord(c) >= 32 or ord(c) in [9, 10, 13]))
 
 
 def text_may_be_spam(text):
-    """
-    >>> text_may_be_spam(u'')
-    False
-    >>> text_may_be_spam(u'  ')
-    False
-    >>> text_may_be_spam(u'this is the content of a blog post')
-    False
-    >>> text_may_be_spam(u'this post contains an http:// link')
-    True
-    >>> text_may_be_spam(u'this post contains an https:// link')
-    True
-    >>> text_may_be_spam(u'this post contains non-basic ascii characters :')
-    True
-    >>> text_may_be_spam(u'this post contains non-basic ascii characters \xc3')
-    True
-    >>> text_may_be_spam(u'this post contains few numbers 1245')
-    False
-    >>> text_may_be_spam(u'this post contains more numbers 1234567')
-    True
-    >>> text_may_be_spam(u'this post contains even more numbers 123456789')
-    True
-    >>> text_may_be_spam(u'this post contains an@email.com')
-    True
-    >>> text_may_be_spam(u'this post contains short.url')
-    True
-    >>> text_may_be_spam(u'BLaCk MaGiC SpEcIaLiSt babaji')
-    True
-    >>> text_may_be_spam(u'love marriage problem solution')
-    True
-    >>> text_may_be_spam(u'fbdad8fbdad8fbdad8')
-    True
-    >>> text_may_be_spam(u'fbdad8')
-    True
+    """Some heuristics to determine if some text may be spam.
+    Arguments:
+        text (basestring): the string to check
+    Returns (bool): True if the text may be spam, False otherwise
     """
 
     # If empty text
@@ -238,15 +169,15 @@ def text_may_be_spam(text):
         return True
 
     # If emails or short urls
-    if re.search("[a-zA-Z0-9\-\.]+\.[a-zA-Z]{2,3}(\s|$|\/|\]|\.)",  text):
+    if re.search(r"[a-zA-Z0-9\-\.]+\.[a-zA-Z]{2,3}(\s|$|\/|\]|\.)",  text):
         return True
 
     # If consecutive numbers
-    if re.search("\(|\)|\d{7}",  text):
+    if re.search(r"\(|\)|\d{7}",  text):
         return True
 
     # If non ascii characters
-    if len(re.sub("[^A-Za-z0-9 ]", "", text, flags=re.UNICODE)) < len(text):
+    if len(re.sub(r"[^A-Za-z0-9 ]", "", text, flags=re.UNICODE)) < len(text):
         return True
 
     # Love, marriage and other everyday topics ;)
@@ -260,11 +191,3 @@ def text_may_be_spam(text):
         return True
 
     return False
-
-
-def _test():
-    import doctest
-    doctest.testmod()
-
-if __name__ == "__main__":
-    _test()
